LPS-60355 Move PassThroughClassLoader to osgi-util to reuse it
