+++ modules/portal/portal-wab-extender/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
 	compile group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	compile group: "com.liferay", name: "com.liferay.osgi.util", version: "2.0.4"
+	compile group: "com.liferay", name: "com.liferay.osgi.util", version: "2.0.5"
 	compile group: "com.liferay", name: "com.liferay.portal.servlet.jsp.compiler", version: "1.0.4"
 	compile group: "com.liferay", name: "org.apache.axis", version: "1.4.LIFERAY-PATCHED-1"
 	compile group: "commons-fileupload", name: "commons-fileupload", transitive: false, version: "1.2.1"
