+++ modules/util/osgi-util/src/main/java/com/liferay/osgi/util/classloader/PassThroughClassLoader.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.wab.extender.internal.adapter;
+package com.liferay.osgi.util.classloader;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -22,10 +22,6 @@ import java.net.URL;
 import java.util.Enumeration;
 
 /**
- * Alters the logic in the <code>isRequestFromVM</code> method of Equinox's
- * <code>BundleLoader</code> class. All of the methods in this class must be
- * implemented even though they just invoke the parent class.
- *
  * @author Carlos Sierra Andrés
  */
 public class PassThroughClassLoader extends ClassLoader {
